requirements.txt
jupyter
pandas-ta
pandas_datareader
matplotlib
